+++ sdk/storage/azure-storage-file-share/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 12.7.0-beta.1 (Unreleased)
 - Fixed a bug that would cause auth failures when building a client by passing an endpoint which had a sas token with protocol set to https,http
+- Fixed a bug where snapshot would be appended to a share snapshot instead of sharesnapshot.
+
 - Added support for the 2020-02-10 service version. 
 - Added support to getFileRanges on a previous snapshot. 
 - Added support to set whether or not smb multichannel is enabled.